Rename url installer package to binary
