+++ pkg/controllers/csi/provisioner/controller_test.go
@@ -20,9 +20,9 @@ import (
 	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/dynatraceclient"
 	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/token"
 	"github.com/Dynatrace/dynatrace-operator/pkg/injection/codemodule/installer"
+	"github.com/Dynatrace/dynatrace-operator/pkg/injection/codemodule/installer/binary"
 	"github.com/Dynatrace/dynatrace-operator/pkg/injection/codemodule/installer/image"
 	"github.com/Dynatrace/dynatrace-operator/pkg/injection/codemodule/installer/job"
-	"github.com/Dynatrace/dynatrace-operator/pkg/injection/codemodule/installer/url"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/installconfig"
 	dtbuildermock "github.com/Dynatrace/dynatrace-operator/test/mocks/pkg/controllers/dynakube/dynatraceclient"
 	installermock "github.com/Dynatrace/dynatrace-operator/test/mocks/pkg/injection/codemodule/installer"
@@ -389,10 +389,10 @@ func createFailingInstaller(t *testing.T) *installermock.Installer {
 	return m
 }
 
-func mockURLInstallerBuilder(t *testing.T, mockedInstaller *installermock.Installer) urlInstallerBuilder {
+func mockURLInstallerBuilder(t *testing.T, mockedInstaller *installermock.Installer) binaryInstallerBuilder {
 	t.Helper()
 
-	return func(_ oneagentclient.APIClient, _ *url.Properties) installer.Installer {
+	return func(_ oneagentclient.APIClient, _ *binary.Properties) installer.Installer {
 		return mockedInstaller
 	}
 }
