+++ pkg/injection/codemodule/installer/binary/installer_test.go
@@ -1,4 +1,4 @@
-package url
+package binary
 
 import (
 	"context"
