+++ pkg/injection/codemodule/installer/binary/unpack.go
@@ -1,4 +1,4 @@
-package url
+package binary
 
 import (
 	"os"
